ENH/REF: add PluginManager.iter_entry_points(), add qiime.sdk.util
Added PluginManager.iter_entry_points() classmethod for getting unloaded entry points (needed for CLI optimizations).

Moved qiime.core.util.parse_type/parse_format/UnknownTypeError to qiime.sdk.util (also accessible from qiime.sdk). This move was also necessary for CLI optimizations. It also avoids circular import errors that previously had to be worked around.
